flyway V20추가 그에 맞춰 엔티티및 dto 수정

+++ src/main/java/com/sist/backend/dto/hotel/RoomResponse.java
@@ -16,7 +16,7 @@ public class RoomResponse {
     private String contentId;
     private String name;
     private Integer capacity;
-    private BigDecimal basePrice;
+    private Integer basePrice;
     private Boolean refundable;
     private Boolean breakfastIncluded;
     private Boolean smoking;
